Bug 1258337 - Refactored PasscodeConfirmViewController into specific VCs instead of relying on confirmAction. Includes added confirmation step when changing passcode.
