+++ Client/Frontend/AuthenticationManager/RemovePasscodeViewController.swift
@@ -1,0 +1,77 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+import Foundation
+import SwiftKeychainWrapper
+import Shared
+
+let NotificationPasscodeDidRemove   = "NotificationPasscodeDidRemove"
+
+/// Displayed to the user when removing a passcode.
+class RemovePasscodeViewController: PagingPasscodeViewController, PasscodeInputViewDelegate {
+    private var confirmCode: String?
+
+    override init() {
+        super.init()
+        self.title = AuthenticationStrings.turnOffPasscode
+        self.panes = [
+            PasscodePane(title: AuthenticationStrings.enterPasscode),
+            PasscodePane(title: AuthenticationStrings.reenterPasscode),
+        ]
+    }
+    
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    override func viewWillAppear(animated: Bool) {
+        super.viewWillAppear(animated)
+        panes.forEach { $0.codeInputView.delegate = self }
+
+        // Don't show the keyboard or allow typing if we're locked out. Also display the error.
+        if authenticationInfo?.isLocked() ?? false {
+            panes.first?.codeInputView.userInteractionEnabled = false
+        } else {
+            panes.first?.codeInputView.becomeFirstResponder()
+        }
+    }
+
+    func passcodeInputView(inputView: PasscodeInputView, didFinishEnteringCode code: String) {
+        switch currentPaneIndex {
+        case 0:
+            // Constraint: Passcode must match current passcode before continuing.
+            if code != authenticationInfo?.passcode {
+                panes[currentPaneIndex].shakePasscode()
+                failIncorrectPasscode(inputView: inputView)
+                return
+            }
+
+            confirmCode = code
+            authenticationInfo?.recordValidation()
+
+            errorToast?.removeFromSuperview()
+            scrollToNextAndSelect()
+        case 1:
+            // Constraint: The first and confirmation codes must match.
+            if confirmCode != code {
+                failMismatchPasscode()
+                resetAllInputFields()
+                scrollToPreviousAndSelect()
+                confirmCode = nil
+                return
+            }
+
+            removePasscode()
+            dismiss()
+        default:
+            break
+        }
+    }
+
+    private func removePasscode() {
+        KeychainWrapper.setAuthenticationInfo(nil)
+        let notificationCenter = NSNotificationCenter.defaultCenter()
+        notificationCenter.postNotificationName(NotificationPasscodeDidRemove, object: nil)
+    }
+}
