+++ UITests/AuthenticationManagerTests.swift
@@ -86,7 +86,48 @@ class AuthenticationManagerTests: KIFTestCase {
         enterPasscodeWithDigits("1337")
         tester().waitForViewWithAccessibilityLabel("Enter a new passcode")
         enterPasscodeWithDigits("2337")
+        tester().waitForViewWithAccessibilityLabel("Re-enter passcode")
+        enterPasscodeWithDigits("2337")
         tester().waitForViewWithAccessibilityLabel("Touch ID & Passcode")
+
+        let info = KeychainWrapper.authenticationInfo()!
+        XCTAssertEqual(info.passcode!, "2337")
+
+        closeAuthenticationManager()
+    }
+
+    func testChangePasscodeShowsErrorStates() {
+        setPasscode("1337", interval: .Immediately)
+
+        openAuthenticationManager()
+        tester().tapViewWithAccessibilityLabel("Change Passcode")
+        tester().waitForViewWithAccessibilityLabel("Enter passcode")
+
+        // Enter wrong passcode
+        enterPasscodeWithDigits("2337")
+        tester().waitForViewWithAccessibilityLabel(String(format: AuthenticationStrings.incorrectAttemptsRemaining, 2))
+
+        enterPasscodeWithDigits("2337")
+        tester().waitForViewWithAccessibilityLabel(String(format: AuthenticationStrings.incorrectAttemptsRemaining, 1))
+
+        enterPasscodeWithDigits("1337")
+        tester().waitForViewWithAccessibilityLabel("Enter a new passcode")
+
+        // Enter same passcode
+        enterPasscodeWithDigits("1337")
+        tester().waitForViewWithAccessibilityLabel("New passcode must be different than existing code.")
+
+        enterPasscodeWithDigits("2337")
+        tester().waitForViewWithAccessibilityLabel("Re-enter passcode")
+
+        // Enter mismatched passcode
+        enterPasscodeWithDigits("3337")
+        tester().waitForViewWithAccessibilityLabel("Passcodes didn't match. Try again.")
+
+        enterPasscodeWithDigits("2337")
+        tester().waitForViewWithAccessibilityLabel("Re-enter passcode")
+
+        enterPasscodeWithDigits("2337")
 
         let info = KeychainWrapper.authenticationInfo()!
         XCTAssertEqual(info.passcode!, "2337")
